Code organization and efficiency
Separated functions into modules, changed keypoint matching to match on grayscale images, combined cropping indexing functions, added check for image pixels in first/last rows/columns of image before finding cropping indices
