+++ .gitignore
@@ -1,0 +1,2 @@
+#Python bytecode cache files
+/modules/__pycache__
